refactoring #101: updated schedule.kpi parser
- refactored schedule.kpi parser
- made parser use campus api
- updated swagger doc

closes: #101
