+++ fictadvisor-api/src/v2/database/entities/DbDisciplineType.ts
@@ -1,0 +1,7 @@
+export class DbDisciplineType {
+  id: string;
+  disciplineId: string;
+  name: string;
+  createdAt: Date;
+  updatedAt: Date;
+}
