+++ fictadvisor-api/src/v2/database/repositories/StudentRepository.ts
@@ -15,7 +15,7 @@ export class StudentRepository {
             speciality: true,
           },
         },
-      }
+      },
     },
     roles: {
       include: {
